Added Visual
This visual is a graph of MIN% which is percentage of total minutes that you played vs. USG% which is basically how involved you are on the court, with the color of the dots based on the VI value, versatility index, which is how versatile you are when it comes to points rebounds and assists in the game.